+++ app-1/src/main/java/com/dxc/entity/Employee.java
@@ -12,7 +12,7 @@ public class Employee {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
-	private Long id;
+	private int id;
 	private String name;
 	private int mobile;
 	private int salary;
@@ -20,18 +20,18 @@ public class Employee {
 	public Employee() {
 	}
 
-	public Employee(Long id, String name, int mobile, int salary) {
+	public Employee(int id, String name, int mobile, int salary) {
 		this.id = id;
 		this.name = name;
 		this.mobile = mobile;
 		this.salary = salary;
 	}
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
